LLVM.Core

Description

The LLVM (Low Level Virtual Machine) is virtual machine at a machine code level. It supports both stand alone code generation and JITing. The Haskell llvm package is a (relatively) high level interface to the LLVM. The high level interface makes it easy to construct LLVM code. There is also an interface to the raw low level LLVM API as exposed by the LLVM C interface.

LLVM code is organized into modules (type Module). Each module contains a number of global variables and functions (type Function). Each functions has a number of basic blocks (type BasicBlock). Each basic block has a number instructions, where each instruction produces a value (type Value).

Unlike assembly code for a real processor the assembly code for LLVM is in SSA (Static Single Assignment) form. This means that each instruction generates a new bound variable which may not be assigned again. A consequence of this is that where control flow joins from several execution paths there has to be a phi pseudo instruction if you want different variables to be joined into one.

The definition of several of the LLVM entities (Module, Function, and BasicBlock) follow the same pattern. First the entity has to be created using newX (where X is one of Module, Function, or BasicBlock), then at some later point it has to given its definition using defineX. The reason for splitting the creation and definition is that you often need to be able to refer to an entity before giving it's body, e.g., in two mutually recursive functions. The the newX and defineX function can also be done at the same time by using createX. Furthermore, an explicit name can be given to an entity by the newNamedX function; the newX function just generates a fresh name.

Initialize

initializeNativeTarget :: IO ()

Modules

data Module

Type of top level modules.

newModule :: IO Module

Create a new module.

newNamedModule

Arguments

:: String
module name

-> IO Module

Create a new explicitely named module.

defineModule

Arguments

:: Module
module that is defined

-> CodeGenModule a
module body

-> IO a

Give the body for a module.

destroyModule :: Module -> IO ()

Free all storage related to a module. *Note*, this is a dangerous call, since referring to the module after this call is an error. The reason for the explicit call to free the module instead of an automatic lifetime management is that modules have a somewhat complicated ownership. Handing a module to a module provider changes the ownership of the module, and the module provider will free the module when necessary.

createModule

Arguments

:: CodeGenModule a
module body

-> IO a

Create a new module with the given body.

data ModuleProvider

A module provider is used by the code generator to get access to a module.

createModuleProviderForExistingModule :: Module -> IO ModuleProvider

Turn a module into a module provider.

data PassManager

Manage compile passes.

createPassManager :: IO PassManager

Create a pass manager.

createFunctionPassManager :: ModuleProvider -> IO PassManager

Create a pass manager for a module.

writeBitcodeToFile :: String -> Module -> IO ()

Write a module to a file.

readBitcodeFromFile :: String -> IO Module

Read a module from a file.

getModuleValues :: Module -> IO [(String, ModuleValue)]

getFunctions :: Module -> IO [(String, Value)]

getGlobalVariables :: Module -> IO [(String, Value)]

data ModuleValue

castModuleValue :: forall a. IsType a => ModuleValue -> Maybe (Value a)

Instructions

ADT representation of IR

data BinOpDesc

Constructors

BOAdd

BOAddNuw

BOAddNsw

BOAddNuwNsw

BOFAdd

BOSub

BOSubNuw

BOSubNsw

BOSubNuwNsw

BOFSub

BOMul

BOMulNuw

BOMulNsw

BOMulNuwNsw

BOFMul

BOUDiv

BOSDiv

BOSDivExact

BOFDiv

BOURem

BOSRem

BOFRem

BOShL

BOLShR

BOAShR

BOAnd

BOOr

BOXor

data InstrDesc

Constructors

IDRet TypeDesc ArgDesc

IDRetVoid

IDBrCond ArgDesc ArgDesc ArgDesc

IDBrUncond ArgDesc

IDSwitch [(ArgDesc, ArgDesc)]

IDIndirectBr

IDInvoke

IDUnwind

IDUnreachable

IDBinOp BinOpDesc TypeDesc ArgDesc ArgDesc

IDAlloca TypeDesc Int Int

IDLoad TypeDesc ArgDesc

IDStore TypeDesc ArgDesc ArgDesc

IDGetElementPtr TypeDesc [ArgDesc]

IDTrunc TypeDesc TypeDesc ArgDesc

IDZExt TypeDesc TypeDesc ArgDesc

IDSExt TypeDesc TypeDesc ArgDesc

IDFPtoUI TypeDesc TypeDesc ArgDesc

IDFPtoSI TypeDesc TypeDesc ArgDesc

IDUItoFP TypeDesc TypeDesc ArgDesc

IDSItoFP TypeDesc TypeDesc ArgDesc

IDFPTrunc TypeDesc TypeDesc ArgDesc

IDFPExt TypeDesc TypeDesc ArgDesc

IDPtrToInt TypeDesc TypeDesc ArgDesc

IDIntToPtr TypeDesc TypeDesc ArgDesc

IDBitcast TypeDesc TypeDesc ArgDesc

IDICmp IntPredicate ArgDesc ArgDesc

IDFCmp FPPredicate ArgDesc ArgDesc

IDPhi TypeDesc [(ArgDesc, ArgDesc)]

IDCall TypeDesc ArgDesc [ArgDesc]

IDSelect TypeDesc ArgDesc ArgDesc

IDUserOp1

IDUserOp2

IDVAArg

IDExtractElement

IDInsertElement

IDShuffleVector

IDExtractValue

IDInsertValue

IDInvalidOp

data ArgDesc

Constructors

AV String

AI Int

AL String

AE

getInstrDesc :: ValueRef -> IO (String, InstrDesc)

Terminator instructions

ret :: Ret a r => a -> CodeGenFunction r Terminate

Return from the current function with the given value. Use () as the return value for what would be a void function is C.

condBr

Arguments

:: Value Bool
Boolean to branch upon.

-> BasicBlock
Target for true.

-> BasicBlock
Target for false.

-> CodeGenFunction r Terminate

Branch to the first basic block if the boolean is true, otherwise to the second basic block.

br

Arguments

:: BasicBlock
Branch target.

-> CodeGenFunction r Terminate

Unconditionally branch to the given basic block.

switch

Arguments

:: IsInteger a

=> Value a
Value to branch upon.

-> BasicBlock
Default branch target.

-> [(ConstValue a, BasicBlock)]
Labels and corresponding branch targets.

-> CodeGenFunction r Terminate

Branch table instruction.

invoke

Arguments

:: CallArgs f g r

=> BasicBlock
Normal return point.

-> BasicBlock
Exception return point.

-> Function f
Function to call.

-> g

Call a function with exception handling.

invokeWithConv

Arguments

:: CallArgs f g r

=> CallingConvention
Calling convention

-> BasicBlock
Normal return point.

-> BasicBlock
Exception return point.

-> Function f
Function to call.

-> g

Call a function with exception handling. This also sets the calling convention of the call to the function. As LLVM itself defines, if the calling conventions of the calling instruction and the function being called are different, undefined behavior results.

unwind :: CodeGenFunction r Terminate

Unwind the call stack until a function call performed with invoke is reached. I.e., throw a non-local exception.

unreachable :: CodeGenFunction r Terminate

Inform the code generator that this code can never be reached.

Arithmetic binary operations

Arithmetic operations with the normal semantics. The u instractions are unsigned, the s instructions are signed.

add :: (IsArithmetic c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

sub :: (IsArithmetic c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

mul :: (IsArithmetic c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

neg :: forall r a. IsArithmetic a => Value a -> CodeGenFunction r (Value a)

iadd :: (IsInteger c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

isub :: (IsInteger c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

imul :: (IsInteger c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

ineg :: IsInteger a => Value a -> CodeGenFunction r (Value a)

fadd :: (IsFloating c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

fsub :: (IsFloating c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

fmul :: (IsFloating c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

fneg :: forall r a. IsFloating a => Value a -> CodeGenFunction r (Value a)

idiv :: forall a b c r v. (IsInteger c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

signed or unsigned integer division depending on the type

irem :: forall a b c r v. (IsInteger c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

signed or unsigned remainder depending on the type

udiv :: (IsInteger c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

sdiv :: (IsInteger c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

fdiv :: (IsFloating c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

Floating point division.

urem :: (IsInteger c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

srem :: (IsInteger c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

frem :: (IsFloating c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

Floating point remainder.

Logical binary operations

Logical instructions with the normal semantics.

shl :: (IsInteger c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

lshr :: (IsInteger c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

ashr :: (IsInteger c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

and :: (IsInteger c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

or :: (IsInteger c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

xor :: (IsInteger c, ABinOp a b (v c)) => a -> b -> CodeGenFunction r (v c)

inv :: IsInteger a => Value a -> CodeGenFunction r (Value a)

Vector operations

extractelement

Arguments

:: Pos n

=> Value (Vector n a)
Vector

-> Value Word32
Index into the vector

-> CodeGenFunction r (Value a)

Get a value from a vector.

insertelement

Arguments

:: Pos n

=> Value (Vector n a)
Vector

-> Value a
Value to insert

-> Value Word32
Index into the vector

-> CodeGenFunction r (Value (Vector n a))

Insert a value into a vector, nondestructive.

shufflevector :: (Pos n, Pos m) => Value (Vector n a) -> Value (Vector n a) -> ConstValue (Vector m Word32) -> CodeGenFunction r (Value (Vector m a))

Permute vector.

Aggregate operation

extractvalue

Arguments

:: forall r agg i a . GetValue agg i a

=> Value agg
Aggregate

-> i
Index into the aggregate

-> CodeGenFunction r (Value a)

Get a value from an aggregate.

insertvalue

Arguments

:: forall r agg i a . GetValue agg i a

=> Value agg
Aggregate

-> Value a
Value to insert

-> i
Index into the aggregate

-> CodeGenFunction r (Value agg)

Insert a value into an aggregate, nondestructive.

Memory access

malloc :: forall a r s. IsSized a s => CodeGenFunction r (Value (Ptr a))

Allocate heap memory.

arrayMalloc :: forall a n r s. (IsSized a n, AllocArg s) => s -> CodeGenFunction r (Value (Ptr a))

Allocate heap (array) memory.

alloca :: forall a r s. IsSized a s => CodeGenFunction r (Value (Ptr a))

Allocate stack memory.

arrayAlloca :: forall a n r s. (IsSized a n, AllocArg s) => s -> CodeGenFunction r (Value (Ptr a))

Allocate stack (array) memory.

free :: IsType a => Value (Ptr a) -> CodeGenFunction r ()

Free heap memory.

load

Arguments

:: Value (Ptr a)
Address to load from.

-> CodeGenFunction r (Value a)

Load a value from memory.

store

Arguments

:: Value a
Value to store.

-> Value (Ptr a)
Address to store to.

-> CodeGenFunction r ()

Store a value in memory

getElementPtr :: forall a o i n r. (GetElementPtr o i n, IsIndexArg a) => Value (Ptr o) -> (a, i) -> CodeGenFunction r (Value (Ptr n))

Address arithmetic. See LLVM description. The index is a nested tuple of the form (i1,(i2,( ... ()))). (This is without a doubt the most confusing LLVM instruction, but the types help.)

getElementPtr0 :: GetElementPtr o i n => Value (Ptr o) -> i -> CodeGenFunction r (Value (Ptr n))

Like getElementPtr, but with an initial index that is 0. This is useful since any pointer first need to be indexed off the pointer, and then into its actual value. This first indexing is often with 0.

Conversions

trunc :: (IsInteger a, IsInteger b, IsPrimitive a, IsPrimitive b, IsSized a sa, IsSized b sb, sa :>: sb) => Value a -> CodeGenFunction r (Value b)

Truncate a value to a shorter bit width.

zext :: (IsInteger a, IsInteger b, IsPrimitive a, IsPrimitive b, IsSized a sa, IsSized b sb, sa :<: sb) => Value a -> CodeGenFunction r (Value b)

Zero extend a value to a wider width.

sext :: (IsInteger a, IsInteger b, IsPrimitive a, IsPrimitive b, IsSized a sa, IsSized b sb, sa :<: sb) => Value a -> CodeGenFunction r (Value b)

Sign extend a value to wider width.

fptrunc :: (IsFloating a, IsFloating b, IsPrimitive a, IsPrimitive b, IsSized a sa, IsSized b sb, sa :>: sb) => Value a -> CodeGenFunction r (Value b)

Truncate a floating point value.

fpext :: (IsFloating a, IsFloating b, IsPrimitive a, IsPrimitive b, IsSized a sa, IsSized b sb, sa :<: sb) => Value a -> CodeGenFunction r (Value b)

Extend a floating point value.

fptoui :: (IsFloating a, IsInteger b, NumberOfElements n a, NumberOfElements n b) => Value a -> CodeGenFunction r (Value b)

Convert a floating point value to an unsigned integer.

fptosi :: (IsFloating a, IsInteger b, NumberOfElements n a, NumberOfElements n b) => Value a -> CodeGenFunction r (Value b)

Convert a floating point value to a signed integer.

fptoint :: forall r n a b. (IsFloating a, IsInteger b, NumberOfElements n a, NumberOfElements n b) => Value a -> CodeGenFunction r (Value b)

Convert a floating point value to an integer. It is mapped to fptosi or fptoui depending on the type a.

uitofp :: (IsInteger a, IsFloating b, NumberOfElements n a, NumberOfElements n b) => Value a -> CodeGenFunction r (Value b)

Convert an unsigned integer to a floating point value.

sitofp :: (IsInteger a, IsFloating b, NumberOfElements n a, NumberOfElements n b) => Value a -> CodeGenFunction r (Value b)

Convert a signed integer to a floating point value.

inttofp :: forall r n a b. (IsInteger a, IsFloating b, NumberOfElements n a, NumberOfElements n b) => Value a -> CodeGenFunction r (Value b)

Convert an integer to a floating point value. It is mapped to sitofp or uitofp depending on the type a.

ptrtoint :: (IsInteger b, IsPrimitive b) => Value (Ptr a) -> CodeGenFunction r (Value b)

Convert a pointer to an integer.

inttoptr :: (IsInteger a, IsType b) => Value a -> CodeGenFunction r (Value (Ptr b))

Convert an integer to a pointer.

bitcast :: (IsFirstClass a, IsFirstClass b, IsSized a sa, IsSized b sb, sa :==: sb) => Value a -> CodeGenFunction r (Value b)

Convert between to values of the same size by just copying the bit pattern.

bitcastUnify :: (IsFirstClass a, IsFirstClass b, IsSized a s, IsSized b s) => Value a -> CodeGenFunction r (Value b)

Same as bitcast but instead of the '(:==:)' type class it uses type unification. This way, properties like reflexivity, symmetry and transitivity are obvious to the Haskell compiler.

Comparison

data CmpPredicate

Constructors

CmpEQ
equal

CmpNE
not equal

CmpGT
greater than

CmpGE
greater or equal

CmpLT
less than

CmpLE
less or equal

data IntPredicate

Constructors

IntEQ
equal

IntNE
not equal

IntUGT
unsigned greater than

IntUGE
unsigned greater or equal

IntULT
unsigned less than

IntULE
unsigned less or equal

IntSGT
signed greater than

IntSGE
signed greater or equal

IntSLT
signed less than

IntSLE
signed less or equal

data FPPredicate

Constructors

FPFalse
Always false (always folded)

FPOEQ
True if ordered and equal

FPOGT
True if ordered and greater than

FPOGE
True if ordered and greater than or equal

FPOLT
True if ordered and less than

FPOLE
True if ordered and less than or equal

FPONE
True if ordered and operands are unequal

FPORD
True if ordered (no nans)

FPUNO
True if unordered: isnan(X) | isnan(Y)

FPUEQ
True if unordered or equal

FPUGT
True if unordered or greater than

FPUGE
True if unordered, greater than, or equal

FPULT
True if unordered or less than

FPULE
True if unordered, less than, or equal

FPUNE
True if unordered or not equal

FPT
Always true (always folded)

class CmpRet c d | c -> d

cmp :: forall a b c d r. (CmpOp a b c d, CmpRet c d) => CmpPredicate -> a -> b -> CodeGenFunction r (Value d)

Compare values of ordered types and choose predicates according to the compared types. Floating point numbers are compared in "ordered" mode, that is NaN operands yields False as result. Pointers are compared unsigned. These choices are consistent with comparison in plain Haskell.

pcmp :: (CmpOp a b (Ptr c) d, CmpRet (Ptr c) d) => IntPredicate -> a -> b -> CodeGenFunction r (Value d)

icmp :: (IsIntegerOrPointer c, CmpOp a b c d, CmpRet c d) => IntPredicate -> a -> b -> CodeGenFunction r (Value d)

Compare integers.

fcmp :: (IsFloating c, CmpOp a b c d, CmpRet c d) => FPPredicate -> a -> b -> CodeGenFunction r (Value d)

Compare floating point values.

select :: (IsFirstClass a, CmpRet a b) => Value b -> Value a -> Value a -> CodeGenFunction r (Value a)

Select between two values depending on a boolean.

Other

phi :: forall a r. IsFirstClass a => [(Value a, BasicBlock)] -> CodeGenFunction r (Value a)

Join several variables (virtual registers) from different basic blocks into one. All of the variables in the list are joined. See also addPhiInputs.

addPhiInputs

Arguments

:: forall a r . IsFirstClass a

=> Value a
Must be a variable from a call to phi.

-> [(Value a, BasicBlock)]
Variables to add.

-> CodeGenFunction r ()

Add additional inputs to an existing phi node. The reason for this instruction is that sometimes the structure of the code makes it impossible to have all variables in scope at the point where you need the phi node.

call :: CallArgs f g r => Function f -> g

Call a function with the given arguments. The call instruction is variadic, i.e., the number of arguments it takes depends on the type of f.

callWithConv :: CallArgs f g r => CallingConvention -> Function f -> g

Call a function with the given arguments. The call instruction is variadic, i.e., the number of arguments it takes depends on the type of f. This also sets the calling convention of the call to the function. As LLVM itself defines, if the calling conventions of the calling instruction and the function being called are different, undefined behavior results.

Classes and types

type Terminate = ()

class Ret a r

Acceptable arguments to the ret instruction.

class CallArgs f g r | g -> r f, f r -> g

Acceptable arguments to call.

class ABinOp a b c | a b -> c

Acceptable arguments to arithmetic binary instructions.

class CmpOp a b c d | a b -> c

Acceptable operands to comparison instructions.

class FunctionArgs f g r | f -> g r, g r -> f

class FunctionArgs (IO a) (CodeGenFunction a ()) a => FunctionRet a

This class is just to simplify contexts.

class IsConst a

class AllocArg a

Acceptable argument to array memory allocation.

class GetElementPtr optr ixs nptr | optr ixs -> nptr

Acceptable arguments to getElementPointer.

class IsIndexArg a

Acceptable single index to getElementPointer.

class GetValue agg ix el | agg ix -> el

Acceptable arguments to extractvalue and insertvalue.

Types classification

Type classifier

class IsType a where

The IsType class classifies all types that have an LLVM representation.

Methods

typeDesc :: a -> TypeDesc

Special type classifiers

class NatI n => Nat n

class PosI n => Pos n

class IsFirstClass a => IsArithmetic a where

Arithmetic types, i.e., integral and floating types.

Methods

arithmeticType :: ArithmeticType a

data ArithmeticType a

Constructors

IntegerType

FloatingType

class (IsArithmetic a, IsIntegerOrPointer a) => IsInteger a

Integral types.

class IsIntegerOrPointer a

Integral or pointer type.

class IsArithmetic a => IsFloating a

Floating types.

class NumberOfElements D1 a => IsPrimitive a

Primitive types.

class IsType a => IsFirstClass a

First class types, i.e., the types that can be passed as arguments, etc.

class (IsType a, Pos s) => IsSized a s | a -> s

Types with a fixed size.

class IsType a => IsFunction a

Function type.

Others

class IsType a => NumberOfElements n a | a -> n

Number of elements for instructions that handle both primitive and vector types

type UnknownSize = D99

Structs

type :& a as = (a, as)

(&) :: a -> as -> a :& as

Type tests

data TypeDesc

Type descriptor, used to convey type information through the LLVM API.

Constructors

TDFloat

TDDouble

TDFP128

TDVoid

TDInt Bool Integer

TDArray Integer TypeDesc

TDVector Integer TypeDesc

TDPtr TypeDesc

TDFunction Bool [TypeDesc] TypeDesc

TDLabel

TDStruct [TypeDesc] Bool

TDInvalidType

isFloating :: IsArithmetic a => a -> Bool

isSigned :: IsInteger a => a -> Bool

typeRef

Arguments

:: IsType a

=> a

-> TypeRef
The argument is never evaluated

typeName :: IsType a => a -> String

typeDesc2 :: TypeRef -> IO TypeDesc

data VarArgs a

The VarArgs type is a placeholder for the real IO type that can be obtained with castVarArgs.

class CastVarArgs a b

Define what vararg types are permissible.

Extra types

newtype IntN n

Variable sized signed integer. The n parameter should belong to PosI.

Constructors

IntN Integer

newtype WordN n

Variable sized unsigned integer. The n parameter should belong to PosI.

Constructors

WordN Integer

newtype FP128

128 bit floating point.

Constructors

FP128 Rational

newtype Array n a

Fixed sized arrays, the array size is encoded in the n parameter.

Constructors

Array [a]

newtype Vector n a

Fixed sized vector, the array size is encoded in the n parameter.

Constructors

Vector [a]

data Ptr a

A value of type Ptr a represents a pointer to an object, or an array of objects, which may be marshalled to or from Haskell values of type a.

The type a will often be an instance of class Storable which provides the marshalling operations. However this is not essential, and you can provide your own operations to access the pointer. For example you might write small foreign functions to get or set the fields of a C struct.

data Label

Label type, produced by a basic block.

newtype Struct a

Struct types; a list (nested tuple) of component types.

Constructors

Struct a

newtype PackedStruct a

Constructors

PackedStruct a

Values and constants

data Value a

data ConstValue a

valueOf :: IsConst a => a -> Value a

value :: ConstValue a -> Value a

zero :: forall a. IsType a => ConstValue a

allOnes :: forall a. IsInteger a => ConstValue a

undef :: forall a. IsType a => ConstValue a

createString :: String -> TGlobal (Array n Word8)

createStringNul :: String -> TGlobal (Array n Word8)

withString :: String -> (forall n. Nat n => Global (Array n Word8) -> CodeGenModule a) -> CodeGenModule a

withStringNul :: String -> (forall n. Nat n => Global (Array n Word8) -> CodeGenModule a) -> CodeGenModule a

constVector :: forall a n. Pos n => [ConstValue a] -> ConstValue (Vector n a)

Make a constant vector. Replicates or truncates the list to get length n.

constArray :: forall a n s. (IsSized a s, Nat n) => [ConstValue a] -> ConstValue (Array n a)

Make a constant array. Replicates or truncates the list to get length n.

constStruct :: IsConstStruct c a => c -> ConstValue (Struct a)

Make a constant struct.

constPackedStruct :: IsConstStruct c a => c -> ConstValue (PackedStruct a)

Make a constant packed struct.

toVector :: MkVector va n a => va -> Vector n a

fromVector :: MkVector va n a => Vector n a -> va

vector :: forall a n. Pos n => [a] -> Vector n a

Make a constant vector. Replicates or truncates the list to get length n. This behaviour is consistent with that of constVector.

Code generation

data CodeGenFunction r a

data CodeGenModule a

Functions

type Function a = Value (Ptr a)

A function is simply a pointer to the function.

newFunction :: forall a. IsFunction a => Linkage -> CodeGenModule (Function a)

Create a new function. Use newNamedFunction to create a function with external linkage, since it needs a known name.

newNamedFunction

Arguments

:: forall a . IsFunction a

=> Linkage

-> String
Function name

-> CodeGenModule (Function a)

Create a new named function.

defineFunction

Arguments

:: forall f g r . FunctionArgs f g r

=> Function f
Function to define (created by newFunction).

-> g
Function body.

-> CodeGenModule ()

Define a function body. The basic block returned by the function is the function entry point.

createFunction

Arguments

:: (IsFunction f, FunctionArgs f g r)

=> Linkage

-> g
Function body.

-> CodeGenModule (Function f)

Create a new function with the given body.

createNamedFunction

Arguments

:: (IsFunction f, FunctionArgs f g r)

=> Linkage

-> String

-> g
Function body.

-> CodeGenModule (Function f)

Create a new function with the given body.

setFuncCallConv :: Function a -> CallingConvention -> CodeGenModule ()

Set the calling convention of a function. By default it is the C calling convention.

type TFunction a = CodeGenModule (Function a)

liftCodeGenModule :: CodeGenModule a -> CodeGenFunction r a

Allows you to define part of a module while in the middle of defining a function.

getParams :: Value -> IO [(String, Value)]

Global variable creation

type Global a = Value (Ptr a)

newGlobal :: forall a. IsType a => Bool -> Linkage -> TGlobal a

Create a new global variable.

newNamedGlobal

Arguments

:: forall a . IsType a

=> Bool
Constant?

-> Linkage
Visibility

-> String
Name

-> TGlobal a

Create a new named global variable.

defineGlobal :: Global a -> ConstValue a -> CodeGenModule ()

Give a global variable a (constant) value.

createGlobal :: IsType a => Bool -> Linkage -> ConstValue a -> TGlobal a

Create and define a global variable.

createNamedGlobal :: IsType a => Bool -> Linkage -> String -> ConstValue a -> TGlobal a

Create and define a named global variable.

externFunction :: forall a r. IsFunction a => String -> CodeGenFunction r (Function a)

Create a reference to an external function while code generating for a function. If LLVM cannot resolve its name, then you may try staticFunction.

staticFunction :: forall f r. IsFunction f => FunPtr f -> CodeGenFunction r (Function f)

Make an external C function with a fixed address callable from LLVM code. This callback function can also be a Haskell function, that was imported like

 foreign import ccall "&nextElement"
    nextElementFunPtr :: FunPtr (StablePtr (IORef [Word32]) -> IO Word32)

See examples/List.hs.

When you only use externFunction, then LLVM cannot resolve the name. (However, I do not know why.) Thus staticFunction manages a list of static functions. This list is automatically installed by simpleFunction and can be manually obtained by getGlobalMappings and installed by addGlobalMappings. "Installing" means calling LLVM's addGlobalMapping according to http://old.nabble.com/jit-with-external-functions-td7769793.html.

externGlobal :: forall a r. IsType a => Bool -> String -> CodeGenFunction r (Global a)

As externFunction, but for Globals rather than Functions

staticGlobal :: forall a r. IsType a => Bool -> Ptr a -> CodeGenFunction r (Global a)

As staticFunction, but for Globals rather than Functions

data GlobalMappings

getGlobalMappings :: CodeGenModule GlobalMappings

Get a list created by calls to staticFunction that must be passed to the execution engine via addGlobalMappings.

type TGlobal a = CodeGenModule (Global a)

Globals

data Linkage

Constructors

ExternalLinkage

AvailableExternallyLinkage

LinkOnceAnyLinkage

LinkOnceODRLinkage

WeakAnyLinkage

WeakODRLinkage

AppendingLinkage

InternalLinkage

PrivateLinkage

DLLImportLinkage

DLLExportLinkage

ExternalWeakLinkage

GhostLinkage

CommonLinkage

LinkerPrivateLinkage

Basic blocks

data BasicBlock

A basic block is a sequence of non-branching instructions, terminated by a control flow instruction.

newBasicBlock :: CodeGenFunction r BasicBlock

newNamedBasicBlock :: String -> CodeGenFunction r BasicBlock

defineBasicBlock :: BasicBlock -> CodeGenFunction r ()

createBasicBlock :: CodeGenFunction r BasicBlock

getCurrentBasicBlock :: CodeGenFunction r BasicBlock

getBasicBlocks :: Value -> IO [(String, Value)]

fromLabel :: Value Label -> BasicBlock

toLabel :: BasicBlock -> Value Label

getInstructions :: Value -> IO [(String, Value)]

getOperands :: Value -> IO [(String, Value)]

hasUsers :: Value -> IO Bool

getUsers :: [Use] -> IO [(String, Value)]

getUses :: Value -> IO [Use]

getUser :: Use -> IO Value

isChildOf :: BasicBlock -> Value -> IO Bool

getDep :: Use -> IO (String, String)

Misc

addAttributes :: Value a -> Int -> [Attribute] -> CodeGenFunction r ()

Add attributes to a value. Beware, what attributes are allowed depends on what kind of value it is.

data Attribute

Constructors

ZExtAttribute

SExtAttribute

NoReturnAttribute

InRegAttribute

StructRetAttribute

NoUnwindAttribute

NoAliasAttribute

ByValAttribute

NestAttribute

ReadNoneAttribute

ReadOnlyAttribute

NoInlineAttribute

AlwaysInlineAttribute

OptimizeForSizeAttribute

StackProtectAttribute

StackProtectReqAttribute

NoCaptureAttribute

NoRedZoneAttribute

NoImplicitFloatAttribute

NakedAttribute

castVarArgs :: CastVarArgs a b => Function a -> Function b

Convert a varargs function to a regular function.

Debugging

dumpValue :: Value a -> IO ()

Print a value.

dumpType :: Value a -> IO ()

Print a type.

getValueName :: Value a -> IO String

Get the name of a Value.

annotateValueList :: [Value] -> IO [(String, Value)]