LLVM.Util.Arithmetic

Documentation

type TValue r a = CodeGenFunction r (Value a)

Synonym for CodeGenFunction r (Value a).

class CmpRet a b => Cmp a b | a -> b where

Methods

cmp :: IntPredicate -> Value a -> Value a -> TValue r b

(%==), (%>=), (%>), (%<=), (%<), (%/=) :: CmpRet a b => TValue r a -> TValue r a -> TValue r b

Comparison functions.

(%&&) :: TValue r Bool -> TValue r Bool -> TValue r Bool

Lazy and.

(%||) :: TValue r Bool -> TValue r Bool -> TValue r Bool

Lazy or.

(?) :: IsFirstClass a => TValue r Bool -> (TValue r a, TValue r a) -> TValue r a

Conditional, returns first element of the pair when condition is true, otherwise second.

(??) :: (IsFirstClass a, CmpRet a b) => TValue r b -> (TValue r a, TValue r a) -> TValue r a

retrn :: Ret (Value a) r => TValue r a -> CodeGenFunction r ()

Return a value from an arithFunction.

set :: TValue r a -> CodeGenFunction r (TValue r a)

Use x <- set $ ... to make a binding.

class ArithFunction a b | a -> b, b -> a

arithFunction :: ArithFunction a b => a -> b

Unlift a function with TValue to have Value arguments.

class (UncurryN a (a1 -> CodeGenFunction r b1), LiftTuple r a1 b, UncurryN a2 (b -> CodeGenFunction r b1)) => UnwrapArgs a a1 b1 b a2 r | a -> a1 b1, a1 b1 -> a, a1 -> b, b -> a1, a2 -> b b1, b b1 -> a2

toArithFunction :: (CallArgs f g r, UnwrapArgs a a1 b1 b g r) => Function f -> a

Lift a function from having Value arguments to having TValue arguments.

recursiveFunction :: (CallArgs a g r0, UnwrapArgs a11 a1 b1 b g r0, FunctionArgs a a2 r1, ArithFunction a3 a2, IsFunction a) => (a11 -> a3) -> CodeGenModule (Function a)

Define a recursive arithFunction, gets passed itself as the first argument.

class CallIntrinsic a